cmd/evm/internal/t8ntool: move tx handling to separate file
