+++ cmd/evm/testdata/30/alloc.json
@@ -1,0 +1,16 @@
+{
+    "0x095e7baea6a6c7c4c2dfeb977efac326af552d87" : {
+        "balance" : "0x0de0b6b3a7640000",
+        "code" : "0x60004960005500",
+        "nonce" : "0x00",
+        "storage" : {
+        }
+    },
+    "0xa94f5374fce5edbc8e2a8697c15331677e6ebf0b" : {
+        "balance" : "0x0de0b6b3a7640000",
+        "code" : "0x",
+        "nonce" : "0x00",
+        "storage" : {
+        }
+    }
+}
