+++ cmd/evm/testdata/30/command.sh
@@ -1,0 +1,1 @@
+/home/wins/.retesteth/default/start.sh --state.fork Cancun --state.reward 0 --state.chainid 1 --input.alloc /home/wins/Ethereum/rettmp/default_Cancun_block1_0xe4e2a3/alloc.json --input.txs /home/wins/Ethereum/rettmp/default_Cancun_block1_0xe4e2a3/txs.rlp --input.env /home/wins/Ethereum/rettmp/default_Cancun_block1_0xe4e2a3/env.json --output.basedir /home/wins/Ethereum/rettmp/default_Cancun_block1_0xe4e2a3 --output.result out.json --output.alloc outAlloc.json --output.errorlog /home/wins/Ethereum/rettmp/default_Cancun_block1_0xe4e2a3/error.json
