+++ cmd/evm/testdata/30/env.json
@@ -1,0 +1,22 @@
+{
+    "currentCoinbase" : "0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba",
+    "currentNumber" : "0x01",
+    "currentTimestamp" : "0x03e8",
+    "currentGasLimit" : "0x1000000000",
+    "previousHash" : "0xe4e2a30b340bec696242b67584264f878600dce98354ae0b6328740fd4ff18da",
+    "currentDataGasUsed" : "0x2000",
+    "parentTimestamp" : "0x00",
+    "parentDifficulty" : "0x00",
+    "parentUncleHash" : "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
+    "currentRandom" : "0x0000000000000000000000000000000000000000000000000000000000020000",
+    "withdrawals" : [
+    ],
+    "parentBaseFee" : "0x08",
+    "parentGasUsed" : "0x00",
+    "parentGasLimit" : "0x1000000000",
+    "parentExcessDataGas" : "0x1000",
+    "parentDataGasUsed" : "0x2000",
+    "blockHashes" : {
+        "0" : "0xe4e2a30b340bec696242b67584264f878600dce98354ae0b6328740fd4ff18da"
+    }
+}
