+++ cmd/evm/testdata/30/error.json
@@ -1,0 +1,1 @@
+rlp: input string too short for common.Address, decoding into (types.Transactions)[0](types.BlobTx).To
